+++ src/local/ss/IntegerValue.java
@@ -1,9 +1,14 @@
 package local.ss;
 
+import java.math.BigInteger;
 import java.util.Hashtable;
+
+/**
+ * Interface for integers that are used ExpressionCalculator.
+ */
 
 public interface IntegerValue {
 
-    public Integer getValue(Hashtable<CellAddress, Object> vars) throws SimpleSpreadSheetException;
+    public BigInteger getValue(Hashtable<CellAddress, Object> vars) throws SimpleSpreadSheetException;
 
 }
